pyproject.toml
[project]
name = "text-to-control"
version = "0.2.0"
description = "Agent-based framework for automated control system design"
readme = "README.md"
authors = [
    { name = "Marek Wadinger", email = "marekwadinger5506@gmail.com" },
    { name = "Adriána Bartková", email = "adrianka.bartkova48@gmail.com" },
]
requires-python = ">=3.12"
dependencies = [
    "control>=0.10.2",
    "logfire>=4.13.2",
    "numpy>=2.3.4",
    "pydantic-ai-slim[google]>=1.0.15",
    "pydantic-settings>=2.11.0",
    "pyomo>=6.9.4",
    "rich>=14.2.0",
    "scipy>=1.16.3",
    "streamlit[auth]>=1.52.2",
    "st-gsheets-connection>=0.0.4",
    "pandas-stubs>=2.3.3.260113",
    "types-requests>=2.32.4.20260107",
]

[dependency-groups]
app = ["fastapi>=0.119.0", "uvicorn>=0.37.0"]
dev = ["commitizen>=4.9.1", "pre-commit>=3.0.0"]
lint = ["ty>=0.0.1a21"]
test = ["pytest>=8.0.0", "pytest-cov>=4.0.0"]
security = ["bandit>=1.7.0"]
format = ["ruff>=0.13.3"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true

[tool.ruff]
line-length = 79
target-version = "py312"

[tool.ruff.lint]
ignore = [
    "E501",
    "E741",
    "E722",
    "N802",
    "N803",
    "F841",
    "E402",
    "E305",
    "PLR1722",
    "T201",
    "PLR2004",
    "F811",
    "E302",
    "N806",
    "ANN001",
    "ANN201",
    "S101",
]
select = ["E", "F", "I", "UP"]
